Fix issue in tensorwrap decorator to have float dtype in Tensors. Fix shaping error in utils.atleast_2d so 1d data is interpreted as a single point that needs a 2d wrapping and not as n points in 1 dimension. Test GPQLearning, but cannot test what the policy converges to without a way to map the results back into the stateaction space.

+++ edge/model/inference/tensorwrap.py
@@ -5,7 +5,7 @@ from decorator import decorator
 
 def ensure_tensor(x):
     if not torch.is_tensor(x):
-        x = torch.tensor(x)
+        x = torch.tensor(x).float()
     return x
